Added support for Tianma TM022HDH26
Some chinese 2.2" SPI TFTs found on eBay are based on an Tianma TM022HDH26 panel and driver by an ILI9341. Support for the ILI9341 was already present (thanks to @cnvogelg) but the Watterott MI0283QT-9A used a 9bit bus, whereas the chinese stuff uses an 8bit bus with a DC pin.
I just added the definition for that kind of screen in fbtft_device.c.
Tested and working on RPi running Occidentalis 0.2 with kernel version 3.6.11+.

Signed-off-by: Kamel Makhloufi <melka@blaste.net>