pcd8544: add support for the almost compatible tls8204.
Later versions of the Nokia 3310 display switched the Philips chip for
a almost compatible Teralan version.  Update the driver to match both.

Change from vertical addressing to horizontal addressing.  The V bit
is used to select other extended command modes instead.

Reset the write pointer on each row.  The display is wider in memory
than on the screen so you either have to pad out to 102 columns or
reset.

Set the address of the first line in memory.  The default causes the
display to be shifted down a few lines.

Only set the temperature coefficient on the PDC8544.

Change magic numbers for the width and height equivalents.

Change the render loop to use a temporary variable and more compiler
and ARM architecture friendly versions.  The multiply should end up
hoisted out of the loop, and the test should turn into a conditional
instruction.